+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [keep a changelog](http://keepachangelog.com) and this pr
 ### Changed
 - Do not use absolute path for `tini` executable in default container entrypoint.
 - Faster validation of JSON object input payloads.
+- Update IAP validation example for Android Publisher v3 API.
 
 ### Fixed
 - Correctly read pagination cursor in notification listings.